Renamed kg_construction folder to components
